requirements.txt
Django==1.6.2
MySQL-python==1.2.5
Pillow==2.4.0
South==0.8.4
Whoosh==2.6.0
argparse==1.2.1
django-admin-timestamps==1.1.1
django-allauth==0.16.1
django-autoslug==1.7.2
django-debug-toolbar==1.1
django-extensions==1.3.3
django-grappelli==2.5.3
django-haystack==2.1.0
django-languages==0.1
django-ratings==0.3.7
django-selectable==0.8.0
django-suit-redactor==0.0.2
lxml==3.3.5
oauthlib==0.6.1
pyisbn==1.0.0
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
